requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
transformers>=4.45.0
torch>=2.4.0
torchvision>=0.19.0
pillow>=10.1.0
pydantic==2.5.0
qwen-vl-utils
accelerate
numpy>=1.24.0
pandas>=2.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
psycopg2-binary>=2.9.0
requests>=2.31.0
email-validator>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
